Move PaymentAccount to payment package.
