+++ src/main/java/io/bisq/api/BisqProxy.java
@@ -3,6 +3,7 @@ package io.bisq.api;
 import com.google.common.base.Strings;
 import com.google.inject.Injector;
 import io.bisq.api.model.*;
+import io.bisq.api.model.payment.PaymentAccountHelper;
 import io.bisq.common.app.DevEnv;
 import io.bisq.common.crypto.KeyRing;
 import io.bisq.common.handlers.ErrorMessageHandler;
@@ -135,7 +136,7 @@ public class BisqProxy {
         return marketList;
     }
 
-    public io.bisq.api.model.PaymentAccount addPaymentAccount(io.bisq.api.model.PaymentAccount account) {
+    public io.bisq.api.model.payment.PaymentAccount addPaymentAccount(io.bisq.api.model.payment.PaymentAccount account) {
         final PaymentAccount paymentAccount = PaymentAccountHelper.toBusinessModel(account);
         user.addPaymentAccount(paymentAccount);
         TradeCurrency singleTradeCurrency = paymentAccount.getSingleTradeCurrency();
