+++ src/main/java/io/bisq/api/model/PaymentAccountList.java
@@ -1,5 +1,7 @@
 package io.bisq.api.model;
 
+import io.bisq.api.model.payment.PaymentAccount;
+
 import java.util.List;
 
 public class PaymentAccountList {
