+++ src/main/java/io/bisq/api/model/payment/PaymentAccount.java
@@ -1,4 +1,4 @@
-package io.bisq.api.model;
+package io.bisq.api.model.payment;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
