+++ src/main/java/io/bisq/api/model/payment/PaymentAccountConverter.java
@@ -2,7 +2,7 @@ package io.bisq.api.model.payment;
 
 import io.bisq.core.payment.PaymentAccount;
 
-public interface PaymentAccountConverter<B extends PaymentAccount, R extends io.bisq.api.model.PaymentAccount> {
+public interface PaymentAccountConverter<B extends PaymentAccount, R extends io.bisq.api.model.payment.PaymentAccount> {
 
     B toBusinessModel(R rest);
