+++ src/main/java/io/bisq/api/model/payment/SepaPaymentAccount.java
@@ -1,7 +1,6 @@
 package io.bisq.api.model.payment;
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
-import io.bisq.api.model.PaymentAccount;
 import io.bisq.api.model.validation.CountryCode;
 import io.bisq.core.payment.payload.PaymentMethod;
 import org.hibernate.validator.constraints.NotBlank;
