+++ src/main/java/io/bisq/api/model/payment/SpecificBanksAccountPaymentAccount.java
@@ -1,7 +1,6 @@
 package io.bisq.api.model.payment;
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
-import io.bisq.api.model.PaymentAccount;
 import io.bisq.core.payment.payload.PaymentMethod;
 import org.hibernate.validator.constraints.NotBlank;
 import org.hibernate.validator.constraints.NotEmpty;
